Added some security procedures to courses view and profile view

+++ courses/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -67,11 +68,15 @@ def course_detail(request, course_id):
 
     return render(request, 'courses/course_detail.html', context)
 
-
+@login_required
 def add_course(request):
     """
     Add a course to the store
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only Store Management can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = CourseForm(request.POST, request.FILES)
         if form.is_valid():
@@ -90,11 +95,15 @@ def add_course(request):
     
     return render(request, template, context)
 
-
+@login_required
 def edit_course(request, course_id):
     """
     Edit a course in the store
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only Store Management can do that.')
+        return redirect(reverse('home'))
+
     course = get_object_or_404(Course, pk=course_id)
     if request.method == 'POST':
         form = CourseForm(request.POST, request.FILES, instance=course)
@@ -116,11 +125,15 @@ def edit_course(request, course_id):
     
     return render(request, template, context)
 
-
+@login_required
 def delete_course(request, course_id):
     """
     Delete a course from store
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only Store Management can do that.')
+        return redirect(reverse('home'))
+        
     course = get_object_or_404(Course, pk=course_id)
     course.delete()
     messages.success(request, 'Course has been deleted!')
